requirements.txt
transformers>=4.42.0
accelerate>=0.30.0
torch>=2.1.0
timm>=0.9.16
pillow
sentencepiece
bitsandbytes
matplotlib
huggingface_hub
python-dotenv
